requirements.txt
pyyaml
requests
six
lxml
